backend/requirements.txt
# Web Server (FastAPI)
fastapi
uvicorn[standard]
pydantic

# AI & LLM 
langchain
langchain-openai
openai
tiktoken

# Sandbox Management (Docker SDK)
docker


# Database
sqlalchemy

# Version Control
GitPython
